% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHpop_map.R
\docType{data}
\name{DHpop_map}
\alias{DHpop_map}
\title{Name of the file containing genotypic (marker scores) information.}
\format{A data frame 130 row (markers) and 3 column.Column 1 gives the marker names,
column 2 the chromosome on which the marker has been mapped, and column 3 indicates
the position of the marker within the chromosome.}
\usage{
DHpop_map
}
\description{
For this example, a doubled haploid population of 200 individuals was
developed by crossing a short line (parent A), with a tall line (parent B).
The 200 doubled haploid lines were evaluated in a field trial and the plant
height measured in cm (the data is in the file DHpop_pheno.txt).
The population was also genotyped by SNPs, and the data is presented in the
file DHpop_geno.txt.
The map location of each of the markers is in the file DHpop_map.txt.
}
\keyword{datasets}

