% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHpop_pheno.R
\docType{data}
\name{DHpop_pheno}
\alias{DHpop_pheno}
\title{Name of the file containing genotypic (marker scores) information.}
\format{A data frame 200 genotypes and 1 variable.}
\usage{
DHpop_pheno
}
\description{
For this example, a doubled haploid population of 200 individuals was
developed by crossing a short line (parent A), with a tall line (parent B).
The 200 doubled haploid lines were evaluated in a field trial and the plant
height measured in cm (the data is in the file DHpop_pheno).
The population was also genotyped by SNPs, and the data is presented in the
file DHpop_geno
The map location of each of the markers is in the file DHpop_map.
}
\keyword{datasets}

